requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
korail2==0.4.0
pycryptodome==3.23.0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
SRTrain==2.6.7
urllib3==2.6.3
